requirements.txt
langchain
langchain-text-splitters
langchain-core
langchain-ollama
langchain-community
langchain-classic
langchain-google-genai
streamlit
yfinance
python-dotenv
chromadb
pypdf
ollama
